Fixed a bug in data range for strings

+++ web/src/main/java/org/hillview/dataStructures/PrivateHeatmap.java
@@ -82,7 +82,7 @@ public class PrivateHeatmap implements Serializable, IJson {
             for (int j = 0; j < this.heatmap.buckets[i].length; j++) {
                 this.noiseForBucket(xIntervals.get(i), yIntervals.get(j), dist, baseVariance, noise);
                 this.heatmap.buckets[i][j] += noise.noise;
-                this.heatmap.confidence[i][j] = 2 * Math.sqrt(noise.variance);
+                this.heatmap.confidence[i][j] = (int)(2 * Math.sqrt(noise.variance));
             }
         }
     }
